requirements.txt
Appium-Python-Client==2.6.2
selene==2.0.b10
pytest==7.1.3
allure-pytest==2.10.0
